build: add ValveFileVDF library to cmake project
