+++ tests/vdf_parser_test.cpp
@@ -6,7 +6,7 @@
 #include <string>
 
 #define TYTI_NO_L_UNDEF
-#include "../vdf_parser.hpp"
+#include "vdf_parser.hpp"
 #define T_L(x) TYTI_L(charT, x)
 using namespace tyti;
